feat : 크리에이터 한줄소개로 변경
